Move print_paf() to public interface
