+++ cudamapper/CMakeLists.txt
@@ -59,6 +59,7 @@ cuda_add_library(cudamapper
         src/cudamapper_utils.cpp
         src/overlapper.cpp
         src/overlapper_triggered.cu
+        src/utils.cpp
         ${CMAKE_CURRENT_BINARY_DIR}/version.cpp)
 
 target_include_directories(cudamapper
