+++ cudamapper/include/claraparabricks/genomeworks/cudamapper/utils.hpp
@@ -1,0 +1,57 @@
+/*
+* Copyright 2019-2020 NVIDIA CORPORATION.
+*
+* Licensed under the Apache License, Version 2.0 (the "License");
+* you may not use this file except in compliance with the License.
+* You may obtain a copy of the License at
+*
+*     http://www.apache.org/licenses/LICENSE-2.0
+*
+* Unless required by applicable law or agreed to in writing, software
+* distributed under the License is distributed on an "AS IS" BASIS,
+* WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+* See the License for the specific language governing permissions and
+* limitations under the License.
+*/
+
+#pragma once
+
+#include <mutex>
+#include <vector>
+#include <string>
+
+#include <claraparabricks/genomeworks/cudamapper/types.hpp>
+
+namespace claraparabricks
+{
+
+namespace genomeworks
+{
+
+namespace io
+{
+class FastaParser;
+}; // namespace io
+
+namespace cudamapper
+{
+
+/// \brief prints overlaps to stdout in <a href="https://github.com/lh3/miniasm/blob/master/PAF.md">PAF format</a>
+/// \param overlaps vector of overlap objects
+/// \param cigars CIGAR strings. Empty vector if none exist
+/// \param query_parser needed for read names and lengths
+/// \param target_parser needed for read names and lengths
+/// \param kmer_size minimizer kmer size
+/// \param write_output_mutex mutex that enables exclusive access to output stream
+void print_paf(const std::vector<Overlap>& overlaps,
+               const std::vector<std::string>& cigars,
+               const io::FastaParser& query_parser,
+               const io::FastaParser& target_parser,
+               int32_t kmer_size,
+               std::mutex& write_output_mutex);
+
+} // namespace cudamapper
+
+} // namespace genomeworks
+
+} // namespace claraparabricks
